MULE-9649: Check that different application plugins do not export the same java packages
_ Validating exported packages on application plugins
